refactor abci server
